fix: Activate external process architecture and update build workflow
CRITICAL FIX: The external process architecture was implemented but never
activated. The build was still using the old in-process HTTP server that
causes BEX64 crashes.

Changes:
1. Activated external process architecture:
   - Renamed plugin.cpp → plugin_old.cpp (old in-process version)
   - Renamed plugin_new.cpp → plugin.cpp (new external process stub)
   - Renamed CMakeLists.txt → CMakeLists_old.txt (old single-target)
   - Renamed CMakeLists_new.txt → CMakeLists.txt (new dual-target)

2. Updated GitHub Actions release workflow:
   - Copy x64dbg_mcp_server.exe to release artifacts
   - Updated installation instructions (all 3 files required)
   - Added external process architecture notes

The new build creates:
- x64dbg_mcp.dp64 / .dp32 (plugin DLL)
- x64dbg_mcp_server.exe (HTTP server)

IMPORTANT: Both files MUST be deployed together in the same directory!

This fixes the persistent BEX64 crash at fault offset 0x0000000300905a4d.
